refactor: move state and methods from App and Quiz components to context for cleaner state management
