+++ src/main.tsx
@@ -4,11 +4,14 @@ import App from "./App.tsx"
 import { ApolloProvider } from "@apollo/client"
 import client from "./apolloClient.ts"
 import { BrowserRouter } from "react-router-dom"
+import { QuizProvider } from "./context/quiz/quiz.tsx"
 
 createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
     <ApolloProvider client={client}>
-      <App />
+      <QuizProvider>
+        <App />
+      </QuizProvider>
     </ApolloProvider>
   </BrowserRouter>
 )
